babyCreation fix (DB-safe)
